restructure acquisition
