--- config/surrogate/acq_fn/botorch.yaml
@@ -1,3 +1,0 @@
-_target_: activelearning.surrogate.acquisition.acquisition.BOTorchAcqHandler
-
-acq_fn_class: qLowerBoundMaxValueEntropy
