--- config/surrogate/surrogate.yaml
@@ -1,6 +1,0 @@
-_target_: activelearning.surrogate.surrogate.SurrogatePipeline
-
-defaults:
-  - _self_
-  - acq_fn: botorch
-  - surrogate_model: gp
