+++ config/test_branin.yaml
@@ -4,7 +4,8 @@ defaults:
   - oracle: branin
   - selector: score
   - sampler: gflownet
-  - surrogate: surrogate
+  - surrogate: gp
+  - acquisition: botorch
   - user: default
